Fix compilation warnings across multiple modules
1. llama-graph-backend-binding.cpp (lines 20-24):
   - Converted C++20 designated initializers to traditional struct initialization
   - Changed .field = value syntax to positional field initialization

2. llama-fallback-elimination.cpp (lines 22-31):
   - Converted C++20 designated initializers to traditional struct initialization
   - Changed .field = value syntax to positional field initialization

3. llama-gpu-memory-fragmentation-monitor.cpp (line 222):
   - Added (void) suppression for unused parameters: buffer_name, size
   - Added comments: "Reserved for future use"

4. llama-pcie-traffic-watchdog.cpp (lines 143, 259):
   - Added static_cast<uint64_t>() around chrono::high_resolution_clock::now().time_since_epoch().count()
   - Fixes narrowing conversion warnings from long int to uint64_t

All changes maintain C++17 compatibility and eliminate compiler warnings.

+++ src/llama-pcie-traffic-watchdog.cpp
@@ -140,7 +140,7 @@ bool pcie_traffic_watchdog::record_transfer(pcie_transfer_type type,
         direction,
         size_bytes,
         source_location,
-        std::chrono::high_resolution_clock::now().time_since_epoch().count(),
+        static_cast<uint64_t>(std::chrono::high_resolution_clock::now().time_since_epoch().count()),
         monitoring_active.load(),
         false  // violation (set below)
     };
@@ -256,7 +256,7 @@ bool pcie_traffic_watchdog::record_violation(const char * description,
         direction,
         size,
         location,
-        std::chrono::high_resolution_clock::now().time_since_epoch().count(),
+        static_cast<uint64_t>(std::chrono::high_resolution_clock::now().time_since_epoch().count()),
         is_critical
     };
